+++ libraries/FastTrig/FastTrig.h
@@ -2,13 +2,11 @@
 //
 //    FILE: FastTrig.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.2.1
 // PURPOSE: Arduino library for a faster approximation of sin() and cos()
 //    DATE: 2011-08-18
 //     URL: https://github.com/RobTillaart/FastTrig
 //          https://forum.arduino.cc/index.php?topic=69723.0
-//
-// HISTORY: see changelog.md
 
 
 #ifdef ESP_PLATFORM
@@ -20,7 +18,7 @@
 #endif
 
 
-#define FAST_TRIG_LIB_VERSION             (F("0.2.0"))
+#define FAST_TRIG_LIB_VERSION             (F("0.2.1"))
 
 #ifdef __cplusplus
 extern "C"
@@ -53,9 +51,17 @@ float iasin(float f);
 
 float iacos(float f);
 
-//  PLACEHOLDER no good implementation (do not use).
+//  PLACEHOLDER (might be obsolete due to atanFast() formula.
 float iatan(float f);
 
+//  fast atan() formula, in fact a modified Taylor expansion
+//  input = -1 .. 1
+float atanFast(float f);
+
+inline float atanHelper(float x);
+
+//  atan2Fast() folds and mirrors => calls atanFast() + offset.
+float atan2Fast(float y, float x);
 
 #ifdef __cplusplus
 }