+++ libraries/FastTrig/test/unit_test_001.cpp
@@ -89,6 +89,14 @@ unittest(test_max_error_table8)
 }
 
 
+unittest(test_atanFast)
+{
+  for (float x = -10; x < 10; x += 0.1)
+  {
+    assertEqualFloat(atanFast(x), atan(x), 0.001);
+  }
+}
+
 unittest_main()
 
 // --------