+++ libraries/FastTrig/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.1] - 2022-12-05
+- add atanFast(x) => faster and less accurate than atan(). 
+  Input range (-1..1) is fastest.
+- add atan2Fast(y, x) => faster and less accurate. 
+  calls atanFast() + offset.
+- add example to measure performance atanFast
+- update readme.md
+- update keywords.txt
+
+
 ## [0.2.0] - 2022-12-02
 - merged PR #16 for ESP_IDF support
 - added <stdint.h>